% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuples.R
\name{all_tuples}
\alias{all_tuples}
\title{Create all tuples of size \code{n}.}
\usage{
all_tuples(tsize, alphabet = nuc_bases())
}
\arguments{
\item{tsize}{Tuple size (e.g. 3 for codons)}

\item{alphabet}{The alphabet. Default are DNA bases.}
}
\description{
Create all tuples of size \code{n}.
}
\examples{
codons = all_tuples(3)
}
