% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuples.R
\name{amino_acids.character}
\alias{amino_acids.character}
\title{Translate codons into amino acids.}
\usage{
\method{amino_acids}{character}(codons, numcode = 1)
}
\arguments{
\item{codons}{Vector of codons.}

\item{numcode}{The ncbi genetic code number for translation.
By default the standard genetic code (1) is used.
See https://www.ncbi.nlm.nih.gov/Taxonomy/Utils/wprintgc.cgi}
}
\description{
Translate codons into amino acids.
}
