% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuples.R
\name{amino_acids.default}
\alias{amino_acids.default}
\title{Default implementation of the generic function for amino acids.
It will throw an stop-exception.}
\usage{
\method{amino_acids}{default}(x, value, ...)
}
\arguments{
\item{x}{Generic parameter}

\item{value}{Generic parameter}

\item{...}{
  Arguments passed on to \code{\link[=amino_acids]{amino_acids}}
  \describe{
    \item{\code{}}{}
  }}
}
\description{
Default implementation of the generic function for amino acids.
It will throw an stop-exception.
}
