% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{amino_acids.gcat.code}
\alias{amino_acids.gcat.code}
\title{Translate codons of a code into amino acids.}
\usage{
\method{amino_acids}{gcat.code}(code, numcode = 1)
}
\arguments{
\item{code}{A code with tuples of size 3.}

\item{numcode}{The ncbi genetic code number for translation.
By default the standard genetic code (1) is used.
See https://www.ncbi.nlm.nih.gov/Taxonomy/Utils/wprintgc.cgi}
}
\description{
Translate codons of a code into amino acids.
}
