% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuples.R
\name{amino_acids}
\alias{amino_acids}
\title{Generic function for amino acids.}
\usage{
amino_acids(x, ...)
}
\arguments{
\item{x}{Generic parameter}

\item{...}{Not used}
}
\description{
Generic function for amino acids.
}
