% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuples.R
\name{as.char.vector}
\alias{as.char.vector}
\title{Convert string into vector of characters.}
\usage{
as.char.vector(s)
}
\arguments{
\item{s}{String, e.g. sequence}
}
\value{
String as vector of characters
}
\description{
Convert string into vector of characters.
}
