% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{classify}
\alias{classify}
\title{Map a sequence to classes "in code" (1) or "out of code" (0).}
\usage{
classify(seq, code)
}
\arguments{
\item{seq}{Sequence as string.}

\item{code}{Code (instance of \code{gcat.code})}
}
\value{
Vector of 0s and 1s.
}
\description{
Map a sequence to classes "in code" (1) or "out of code" (0).
}
