% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{code}
\alias{code}
\title{Create a code from a list of tuples.}
\usage{
code(tuples, id = DEFAULT_ID, unique.set = T, sorted.set = T)
}
\arguments{
\item{tuples}{Vector of tuples or codons as strings, e.g. c("AUC", "GCA").}

\item{id}{A brief description of the code. Default is "unkn.".}

\item{unique.set}{A boulian value. If true, all dublicates are removed from the set.}

\item{sorted.set}{A boulian value. If true, the set is sorted in alphabetical order.}
}
\value{
A 'gcat.code' object. Derived from vector with two attributs: 1) id,  2) tsize.
}
\description{
Create a code from a list of tuples.
}
