% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{get.id}
\alias{get.id}
\title{Returns the id of a code or the default id ("unkn.").}
\usage{
get.id(code)
}
\arguments{
\item{code}{A 'gcat.code' object}
}
\value{
The id of a 'gcat.code'
}
\description{
Returns the id of a code or the default id ("unkn.").
}
\examples{
get.id(code(c("ACG", "CC"), id = "Test 1"))
}
