% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuples.R
\name{normalize}
\alias{normalize}
\title{Normalize a nucleotide sequence. Default is DNA bases and upper-case letters.}
\usage{
normalize(seqv, RNA = FALSE, lowercase = FALSE)
}
\arguments{
\item{seqv}{A vector of sequences as a string to be processed.}

\item{RNA}{If true, RNA bases are used, i.e. T becomes U.
If false (default), DNA bases are used, i.e. U becomes T.}

\item{lowercase}{If true, all letters are converted to lowercase (ATG -> atg).
IF false (default), all letters will be uppercase (atg -> ATG)}
}
\value{
Normalized string
}
\description{
Note that unknown bases (e.g. letter K) do not raise any exceptions.
}
\examples{
normalize("auggcc") # will yield ATGGCC.
}
