% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphabet.R
\name{print.gcat.alphabet}
\alias{print.gcat.alphabet}
\title{Pretty prints an alphabet.}
\usage{
\method{print}{gcat.alphabet}(x, ...)
}
\arguments{
\item{x}{The alphabet}

\item{...}{
  Arguments passed on to \code{\link[base:print]{base::print}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
The formatted alphabet
}
\description{
Pretty prints an alphabet.
}
