% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{print.gcat.code}
\alias{print.gcat.code}
\title{Pretty prints a code.}
\usage{
\method{print}{gcat.code}(x, ...)
}
\arguments{
\item{x}{The code}

\item{...}{
  Arguments passed on to \code{\link[base:print]{base::print}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
The formatted code
}
\description{
Pretty prints a code.
}
