% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{random_code}
\alias{random_code}
\title{Random code.}
\usage{
random_code(size = 20, tsize = 3, id = "unkn. rnd")
}
\arguments{
\item{size}{Number of codons in the code. Default is 20.}

\item{tsize}{Number of bases per tuple. Default is 3.}

\item{id}{Brief description of the code. Default is "unkn. rnd.".}
}
\value{
Vector or string based tuples.
}
\description{
Random code.
}
