% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{read_codes}
\alias{read_codes}
\title{Read one or more codes from file.}
\usage{
read_codes(filename, tsize = 3, skip_codeid_col = TRUE)
}
\arguments{
\item{filename}{Textfile with codes in it.}

\item{tsize}{The size of the tuples in the code. Default is 3.}

\item{skip_codeid_col}{If true the first column is left out.}
}
\value{
A list of codes.
}
\description{
The structure of the file is: \cr
}
\section{header with short description \cr}{
1, ACA, ACT, ... \cr
2, ACT, ATT, ... \cr
}

