% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manip.R
\name{rev_compl}
\alias{rev_compl}
\title{Reversed and complementary tuples.}
\usage{
rev_compl(tuples)
}
\arguments{
\item{tuples}{Vector of tuples. Note: tuples are not required
to have the same tuple size.}
}
\value{
Vector of same lengths like tuples where bases are in reverse order
and complementary.
}
\description{
Example: (ATTGA, GCG) -> (TCAAT, CGC)
}
