% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manip.R
\name{split}
\alias{split}
\title{Split string sequence into tupels (codons).}
\usage{
split(seq, tsize = 3, sep = NULL, regexp.sep = NULL)
}
\arguments{
\item{seq}{Sequence as a string.}

\item{tsize}{Tuple size. Default: 3 for codons.}

\item{sep}{Separator, e.g. ","}

\item{regexp.sep}{Separator as regular expresion, e.g. "\link{, }+"}
}
\value{
Vector of tuples. Each tuple is a string.
}
\description{
A sequence can be split either by passing the parameter \code{tsize} or the
parameter \code{sep}. If \code{sep} is set, \code{tsize} is ignored.
}
\details{
In case of \code{tsize} only tuples of exact this size are returned.
If the last tuple has less than \code{tsize} bases it is skipped.

On the other hand, if \code{set} is used, all tuples are returned. There
is also no check if the tuples have the same size.
}
