% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{summary.gcat.code}
\alias{summary.gcat.code}
\title{Give a summary for a code.}
\usage{
\method{summary}{gcat.code}(object, ...)
}
\arguments{
\item{object}{The code}

\item{...}{
  Arguments passed on to \code{\link[base:summary]{base::summary}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
The summary
}
\description{
Give a summary for a code.
}
