% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manip.R
\name{truncate}
\alias{truncate}
\title{Truncate a string at the beginning.}
\usage{
truncate(seq, k)
}
\arguments{
\item{k}{Characters to truncate: 0, 1, 2, etc.}
}
\value{
String or sequence where k nucleotides are truncated at the beginning.
}
\description{
This is useful for a frame shift.
}
