% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{tsize}
\alias{tsize}
\title{The size or sizes of a code's tuples.}
\usage{
tsize(code)
}
\arguments{
\item{code}{A 'gcat.code' object}
}
\value{
The id of a 'gcat.code'
}
\description{
The size or sizes of a code's tuples.
}
\examples{
tsize(code(c("ACG", "CC"))
}
