% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuples.R
\name{tuples_freq}
\alias{tuples_freq}
\title{Frequencies of tuples (codon).}
\usage{
tuples_freq(tuples)
}
\arguments{
\item{tuples}{Vector of tuples (codons)}
}
\value{
Data frame with two columns:
\enumerate{
\item tuple, 2. number of tuples in vector
}
}
\description{
Calculates the absolute frequencies of tuples in a vector.
}
