% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{write_codes}
\alias{write_codes}
\title{Write one or more codes to a file.}
\usage{
write_codes(filename, codes, header = "# codes")
}
\arguments{
\item{filename}{Textfile with codes in it.}

\item{codes}{List of codes.}

\item{header}{Header of text file. Default is "# codes".}
}
\description{
Details see \code{read.codes}.
}
