% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphabet.R
\name{alphabet}
\alias{alphabet}
\title{Get the alphabet defined by the string of vector of strings.}
\usage{
alphabet(s)
}
\value{
List of letters.
}
\description{
Get the alphabet defined by the string of vector of strings.
}
