% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{amino_acids.gcat.code}
\alias{amino_acids.gcat.code}
\title{Translate codons of a code into amino acids.}
\usage{
\method{amino_acids}{gcat.code}(x, numcode = 1, ...)
}
\arguments{
\item{x}{A code with tuples of size 3.}

\item{numcode}{The ncbi genetic code number for translation.
By default the standard genetic code (1) is used.
See https://www.ncbi.nlm.nih.gov/Taxonomy/Utils/wprintgc.cgi}

\item{...}{
  Arguments passed on to \code{\link[=amino_acids]{amino_acids}}
  \describe{
    \item{\code{}}{}
  }}
}
\description{
Translate codons of a code into amino acids.
}
