% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphabet.R
\name{print.gcat.alphabet}
\alias{print.gcat.alphabet}
\title{Pretty prints an alphabet.}
\usage{
\method{print}{gcat.alphabet}(alphabet)
}
\arguments{
\item{alphabet}{The alphabet}
}
\value{

}
\description{
Pretty prints an alphabet.
}
